+++ lib/sidekiq/failures.rb
@@ -1,4 +1,9 @@
-require "sidekiq/web"
+begin
+  require "sidekiq/web"
+rescue LoadError
+  # client-only usage
+end
+
 require "sidekiq/failures/version"
 require "sidekiq/failures/middleware"
 require "sidekiq/failures/web_extension"
@@ -26,21 +31,41 @@ module Sidekiq
     @failures_default_mode || :all
   end
 
+  # Sets the maximum number of failures to track
+  #
+  # If the number of failures exceeds this number the list will be trimmed (oldest
+  # failures will be purged).
+  #
+  # Defaults to 1000
+  # Set to false to disable rotation
+  def self.failures_max_count=(value)
+    @failures_max_count = value
+  end
+
+  # Fetches the failures max count value
+  def self.failures_max_count
+    return 1000 if @failures_max_count.nil?
+
+    @failures_max_count
+  end
+
   module Failures
   end
-end
-
-Sidekiq::Web.register Sidekiq::Failures::WebExtension
-
-if Sidekiq::Web.tabs.is_a?(Array)
-  # For sidekiq < 2.5
-  Sidekiq::Web.tabs << "failures"
-else
-  Sidekiq::Web.tabs["Failures"] = "failures"
 end
 
 Sidekiq.configure_server do |config|
   config.server_middleware do |chain|
     chain.add Sidekiq::Failures::Middleware
+  end
+end
+
+if defined?(Sidekiq::Web)
+  Sidekiq::Web.register Sidekiq::Failures::WebExtension
+
+  if Sidekiq::Web.tabs.is_a?(Array)
+    # For sidekiq < 2.5
+    Sidekiq::Web.tabs << "failures"
+  else
+    Sidekiq::Web.tabs["Failures"] = "failures"
   end
 end